Add square and pot flags to control texture resizing
Textures which are compressed by PVRTexTool are forced to square and power-of-two. Unfortunately, there is no way to override this behaviour. This commit introduces 2 new optional flags to address this issue: `--square` and `--pot`. Both flags accept a parameter to specify whether the texture should be resized smaller (`-`), larger (`+`) or no change at all (`no`). Default parameter is `+` for both flags due to backward compatibility.